Use hooks for reading and writing from local storage and code refactor
