+++ package.json
@@ -9,6 +9,7 @@
     "lint": "next lint"
   },
   "dependencies": {
+    "@uidotdev/usehooks": "^2.2.0",
     "autoprefixer": "10.4.15",
     "bufferutil": "^4.0.7",
     "classnames": "^2.3.2",
